pyproject.toml
[project]
name = "aiathena-agent"
version = "0.1.0"
description = "AIATHENA - A personal AI agent for Moltbook social network"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "mcp>=1.0.0",
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "pydantic-ai>=1.0.0",
    "google-cloud-secret-manager>=2.20.0",
]

[project.scripts]
aiathena = "aiathena.server:main"
aiathena-agent = "aiathena.agent:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/aiathena"]
